# src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Its a REST API Documentation for school app backend",
    "version": "1.0.0",
    "title": "SONET School App Documentation"
  },
  "host": "",
  "basePath": "/",
...
  "schemes": ["http"],
  "components": {
    "schemas": {
...
      "respObj": {
        "type": "object",
        "properties": {
          "result": {
            "type": "string",
            "default": "OK"
          },
          "resultCode": {
            "type": "integer",
            "default": 200
          },
          "message": {
            "type": "string",
            "default": "Query Success"
          },
          "ValidationErrors": {
            "type": "any",
            "default": ""
          },
          "resultTotal": {
            "type": "integer",
            "default": 0
          },
          "totalRows": {
            "type": "integer",
            "default": 0
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/subjects/createAssign": {
      "post": {
        "tags": ["Manage Subjects"],
        "summary": "Create subject/s ",
        "description": "Create subjects and if needed assign it to the classes or to a classs",
        "operationId": "createAssignSubjects",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/tokens/x-access-token"
          },
          {
            "$ref": "#/tokens/refreshToken"
          },
          {
            "name": "createAssignSubjectsBody",
            "description": "createAssignSubjects request body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/manageSubjects/createAssign"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "UnAuthorized user"
          },
          "200": {
            "description": "Successfully created and assigned subjects",
            "content": "application/json",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/respObj"
                },
                {
                  "$ref": "#/components/schemas/manageSubjects/getCreateAssignedSubjects"
                }
              ]
            }
          }
        }
      }
    },
    "/api/subjects/assign": {
      "post": {
        "tags": ["Manage Subjects"],
        "summary": "Assign subject/s ",
        "description": "Assign subjects",
        "operationId": "assignSubjects",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/tokens/x-access-token"
          },
          {
            "$ref": "#/tokens/refreshToken"
          },
          {
            "name": "assignSubjectsBody",
            "description": "assignSubjects request body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/manageSubjects/assignSubjects"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "UnAuthorized user"
          },
          "200": {
            "description": "Successfully created and assigned subjects",
            "content": "application/json",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/respObj"
                },
                {
                  "$ref": "#/components/schemas/manageSubjects/assignedSubjects"
                }
              ]
            }
          }
        }
      }
    },
    "/api/subjects/update": {
      "patch": {
        "tags": ["Manage Subjects"],
        "summary": "Edit subject/s ",
        "description": "Edit subjects",
        "operationId": "updateSubjects",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/tokens/x-access-token"
          },
          {
            "$ref": "#/tokens/refreshToken"
          },
          {
            "name": "Edit Subjects Body",
            "description": "Edit Subjects request body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/manageSubjects/updateSubjects"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "UnAuthorized user"
          },
          "200": {
            "description": "Successfully created and assigned subjects",
            "content": "application/json",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/respObj"
                },
                {
                  "$ref": "#/components/schemas/manageSubjects/getSubjects"
                }
              ]
            }
          }
        }
      }
    },
    "/api/subjects/{subjectId}": {
      "delete": {
        "tags": ["Manage Subjects"],
        "summary": "Delete subject ",
        "description": "Delete subject",
        "operationId": "deleteSubjects",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/tokens/x-access-token"
          },
          {
            "$ref": "#/tokens/refreshToken"
          },
          {
            "name": "subjectId",
            "description": "Give SubjectId to delete Subject",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "403": {
            "description": "UnAuthorized user"
          },
          "200": {
            "description": "Successfully deleted and unassugned the subject",
            "content": "application/json",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/respObj"
                },
                {
                  "$ref": "#/components/schemas/manageSubjects/getSubjects"
                }
              ]
            }
          }
        }
      }
    },
...
  },
...
}